refactor: split CSS into per-domain files with @layer, extract animation.ts
Splits the monolithic client/style.css into domain files under client/css/
(base, buttons, landing, dialog, stars, responsive) with explicit @layer
ordering, and extracts the shooting-stars choreography engine out of
script.ts into its own animation.ts module. Updates CLAUDE.md accordingly.
