+++ client/css/base.css
@@ -1,0 +1,217 @@
+@layer base {
+    /** Main structure */
+    :root {
+        --color-void: #0a0612;
+        --color-void-rgb: 10, 6, 18;
+        --color-chrome: #e6e6f0;
+        --color-magenta: #ff2ec4;
+        --color-magenta-rgb: 255, 46, 196;
+        --color-magenta-light: #ff6ad5;
+        --color-cyan: #00e5ff;
+        --color-cyan-rgb: 0, 229, 255;
+        --color-purple: #7b2ff7;
+    }
+
+    html {
+        overflow: hidden;
+    }
+
+    body {
+        width: 100%;
+        height: 100%;
+        background-color: black;
+    }
+
+    /* CSS-only starfield, sits between the video (z-index -100) and everything
+       else — visible behind the idle console, faded out in sync with #landing
+       once the animation launches (see script.ts) since the video covers the
+       screen from that point on anyway */
+    #starfield {
+        position: fixed;
+        inset: 0;
+        z-index: -99;
+        background-color: var(--color-void);
+        background-image:
+            radial-gradient(
+                1.5px 1.5px at 20px 30px,
+                var(--color-chrome),
+                transparent
+            ),
+            radial-gradient(
+                1px 1px at 90px 120px,
+                var(--color-cyan),
+                transparent
+            ),
+            radial-gradient(
+                1.5px 1.5px at 160px 60px,
+                var(--color-chrome),
+                transparent
+            ),
+            radial-gradient(
+                1px 1px at 230px 180px,
+                var(--color-magenta-light),
+                transparent
+            ),
+            radial-gradient(
+                1.5px 1.5px at 300px 20px,
+                var(--color-chrome),
+                transparent
+            ),
+            radial-gradient(
+                1px 1px at 50px 200px,
+                var(--color-cyan),
+                transparent
+            ),
+            radial-gradient(
+                1px 1px at 270px 140px,
+                var(--color-chrome),
+                transparent
+            );
+        background-repeat: repeat;
+        background-size: 340px 220px;
+        animation:
+            star-drift 120s linear infinite,
+            star-twinkle 3s ease-in-out infinite alternate;
+        transition: opacity 0.6s ease;
+
+        /* animation: none is required here — the twinkle/drift keyframes keep
+           driving opacity forever, which otherwise wins over this rule's own
+           opacity: 0 every other frame instead of actually fading out */
+        &.fade-out {
+            opacity: 0;
+            animation: none;
+        }
+    }
+
+    @keyframes star-drift {
+        from {
+            background-position: 0 0;
+        }
+        to {
+            background-position: -340px 220px;
+        }
+    }
+
+    @keyframes star-twinkle {
+        from {
+            opacity: 0.75;
+        }
+        to {
+            opacity: 1;
+        }
+    }
+
+    video#video {
+        position: fixed;
+        top: 50%;
+        left: 50%;
+        min-width: 100%;
+        min-height: 100%;
+        width: auto;
+        height: auto;
+        z-index: -100;
+        transform: translateX(-50%) translateY(-50%);
+    }
+
+    div#pictures-container {
+        position: fixed;
+        top: 0;
+        left: 0;
+        width: 100%;
+        height: 100%;
+        /* purely decorative, never interactive — without this it sits on top
+           of (and blocks clicks/hover on) #tap-to-play in the stacking order,
+           since it comes later in the DOM and neither has a z-index */
+        pointer-events: none;
+        /* keyframe translate()/scale() distances below are tuned in px for
+           desktop widths; this scales the whole rendered transform down
+           proportionally on narrow viewports so travel stays on-screen and
+           doesn't read as faster than on desktop */
+        --travel-scale: clamp(0.25, calc(100vw / 1400px), 1);
+
+        img {
+            position: absolute;
+            /* box grown by the inverse of --travel-scale so the `scale`
+               below cancels it out at rest — the image stays the same
+               fraction of the viewport as on desktop, only the travel
+               distance actually shrinks */
+            max-width: calc(30% / var(--travel-scale));
+            max-height: calc(30% / var(--travel-scale));
+            top: 0;
+            left: 0;
+            right: 0;
+            bottom: 0;
+            margin: auto;
+            scale: var(--travel-scale);
+
+            &.hide {
+                display: none;
+            }
+        }
+    }
+
+    footer {
+        color: var(--color-chrome);
+        font-family: ui-monospace, "SFMono-Regular", Consolas, monospace;
+        font-size: 12px;
+        position: fixed;
+        bottom: 0;
+        left: 0;
+        right: 0;
+        z-index: 1;
+        text-align: center;
+        padding: 8px 16px;
+        background: linear-gradient(
+            180deg,
+            rgba(var(--color-void-rgb), 0.2),
+            rgba(var(--color-void-rgb), 0.85)
+        );
+        border-top: 1px solid rgba(var(--color-cyan-rgb), 0.35);
+
+        p {
+            margin: 0;
+        }
+
+        .status-dot {
+            display: inline-block;
+            width: 7px;
+            height: 7px;
+            margin-right: 4px;
+            border-radius: 50%;
+            background: var(--color-cyan);
+            box-shadow: 0 0 8px rgba(var(--color-cyan-rgb), 0.9);
+            animation: twinkle 2.4s ease-in-out infinite;
+        }
+
+        .cursor {
+            animation: cursor-blink 1s step-end infinite;
+        }
+
+        a {
+            color: var(--color-magenta-light);
+            text-decoration: underline;
+            text-underline-offset: 2px;
+
+            &:hover {
+                color: var(--color-cyan);
+            }
+
+            &:focus-visible {
+                outline: 2px solid var(--color-cyan);
+                outline-offset: 2px;
+                border-radius: 2px;
+            }
+        }
+    }
+
+    @keyframes cursor-blink {
+        0%,
+        50% {
+            opacity: 1;
+        }
+        51%,
+        100% {
+            opacity: 0;
+        }
+    }
+}
