+++ client/css/buttons.css
@@ -1,0 +1,232 @@
+@layer components {
+    /* shared chrome for both landing CTAs (launch + upload) — same size/shape
+       so neither reads as the "secondary, corner" option anymore */
+    .console-btn {
+        display: inline-flex;
+        align-items: center;
+        gap: 10px;
+        position: relative;
+        clip-path: polygon(
+            14px 0,
+            100% 0,
+            100% calc(100% - 14px),
+            calc(100% - 14px) 100%,
+            0 100%,
+            0 14px
+        );
+        border: 1.5px solid rgba(var(--color-cyan-rgb), 0.5);
+        padding: 20px 36px;
+        background: rgba(var(--color-void-rgb), 0.6);
+        backdrop-filter: blur(14px);
+        -webkit-backdrop-filter: blur(14px);
+        box-shadow: 0 12px 32px rgba(0, 0, 0, 0.45);
+        cursor: pointer;
+        opacity: 1;
+        transition:
+            opacity 0.4s ease,
+            transform 0.2s ease,
+            box-shadow 0.2s ease,
+            background 0.25s ease,
+            color 0.25s ease,
+            border-color 0.25s ease;
+
+        color: var(--color-chrome);
+        font-family: sans-serif;
+        font-size: 1.6em;
+        font-weight: 800;
+        text-align: center;
+        text-transform: uppercase;
+        letter-spacing: 0.03em;
+
+        p {
+            margin: 0;
+        }
+
+        &:hover {
+            transform: translateY(-2px);
+        }
+
+        &:active {
+            transform: translateY(0);
+            filter: brightness(0.95);
+        }
+
+        /* no lift here (unlike :hover/:active) — keyboard focus shouldn't
+           shift the button's position, just call it out with the outline */
+        &:focus-visible {
+            outline: 2px solid var(--color-cyan);
+            outline-offset: 4px;
+        }
+    }
+
+    /* launch — the hero action, magenta + a slow glow pulse */
+    .console-btn--primary {
+        border-color: rgba(var(--color-magenta-rgb), 0.6);
+        box-shadow:
+            0 12px 32px rgba(0, 0, 0, 0.45),
+            0 0 24px rgba(var(--color-magenta-rgb), 0.18);
+        animation: pulse-glow 2.4s ease-in-out infinite;
+
+        /* the sparkles either side of the text twinkle independently, offset
+           so they don't pulse in sync */
+        .spark {
+            display: inline-block;
+            animation: twinkle 2.4s ease-in-out infinite;
+
+            &:last-child {
+                animation-delay: 1.2s;
+            }
+        }
+
+        /* pausing the glow pulse here (rather than fighting it every frame)
+           lets these static box-shadow/transform values actually stick while
+           interacting */
+        &:hover,
+        &:active {
+            animation-play-state: paused;
+        }
+
+        /* gradient fill + void text is the "you're hovering this" signal on
+           top of the lift — .spark inherits `color` from here too, so the
+           sparkles flip to void along with the text */
+        &:hover {
+            background: linear-gradient(
+                135deg,
+                var(--color-magenta-light),
+                var(--color-magenta)
+            );
+            color: var(--color-void);
+            border-color: rgba(var(--color-magenta-rgb), 0.9);
+            box-shadow:
+                0 16px 40px rgba(0, 0, 0, 0.5),
+                0 0 36px rgba(var(--color-magenta-rgb), 0.4);
+        }
+    }
+
+    /* upload — the secondary action, cyan, calmer (no infinite pulse) so the
+       two CTAs read as equally prominent without competing for attention */
+    .console-btn--secondary {
+        &:hover {
+            background: linear-gradient(
+                135deg,
+                var(--color-cyan),
+                var(--color-purple)
+            );
+            color: var(--color-void);
+            border-color: rgba(var(--color-cyan-rgb), 0.9);
+            box-shadow:
+                0 16px 40px rgba(0, 0, 0, 0.5),
+                0 0 36px rgba(var(--color-cyan-rgb), 0.4);
+        }
+    }
+
+    @keyframes pulse-glow {
+        0%,
+        100% {
+            box-shadow:
+                0 12px 32px rgba(0, 0, 0, 0.45),
+                0 0 24px rgba(var(--color-magenta-rgb), 0.18);
+        }
+        50% {
+            box-shadow:
+                0 12px 32px rgba(0, 0, 0, 0.45),
+                0 0 40px rgba(var(--color-magenta-rgb), 0.4);
+        }
+    }
+
+    @keyframes twinkle {
+        0%,
+        100% {
+            opacity: 0.5;
+            transform: scale(0.85) rotate(0deg);
+        }
+        50% {
+            opacity: 1;
+            transform: scale(1.2) rotate(180deg);
+        }
+    }
+
+    /* persistent controls — stays visible/reachable while the animation is
+       running, unlike #landing (see script.ts) */
+    #quick-actions {
+        position: fixed;
+        top: 20px;
+        left: 20px;
+        z-index: 3;
+
+        display: flex;
+        gap: 8px;
+    }
+
+    .dock-btn {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        width: 44px;
+        height: 44px;
+        font-size: 18px;
+        color: var(--color-chrome);
+        background: rgba(var(--color-void-rgb), 0.6);
+        border: 1.5px solid rgba(var(--color-cyan-rgb), 0.5);
+        clip-path: polygon(
+            8px 0,
+            100% 0,
+            100% calc(100% - 8px),
+            calc(100% - 8px) 100%,
+            0 100%,
+            0 8px
+        );
+        backdrop-filter: blur(10px);
+        -webkit-backdrop-filter: blur(10px);
+        cursor: pointer;
+        transition:
+            transform 0.15s ease,
+            box-shadow 0.15s ease,
+            opacity 0.15s ease;
+
+        &:hover {
+            transform: translateY(-2px);
+            box-shadow: 0 0 20px rgba(var(--color-cyan-rgb), 0.4);
+            border-color: rgba(var(--color-cyan-rgb), 0.9);
+        }
+
+        &:active {
+            transform: translateY(0);
+        }
+
+        &:focus-visible {
+            outline: 2px solid var(--color-cyan);
+            outline-offset: 3px;
+        }
+
+        &:disabled {
+            opacity: 0.35;
+            cursor: not-allowed;
+
+            &:hover {
+                transform: none;
+                box-shadow: none;
+                border-color: rgba(var(--color-cyan-rgb), 0.5);
+            }
+        }
+    }
+
+    input#file-upload {
+        opacity: 0;
+        position: absolute;
+        z-index: -1;
+    }
+
+    /* the hidden file input can't be an adjacent sibling of both labels that
+       trigger it (they live in different parts of the DOM), so script.ts
+       toggles this class on <body> on focus/blur instead of relying on
+       :has() — keeps browser support universal and scopes the ring to only
+       the two actual upload triggers, not every .console-btn */
+    body.file-upload-focused {
+        .console-btn--secondary,
+        .dock-btn[for="file-upload"] {
+            outline: 2px solid var(--color-cyan);
+            outline-offset: 4px;
+        }
+    }
+}
